Move all remaining Input manager stuffs to the Ryujinx.Input project
