--- Ryujinx/Input/KeyboardHotkeyState.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Ryujinx.Input
-{
-    [Flags]
-    public enum KeyboardHotkeyState
-    {
-        None,
-        ToggleVSync,
-    }
-}
